requirements.txt
catboost==1.0.6
cycler==0.11.0
distlib==0.3.6
dnspython==2.2.1
et-xmlfile==1.1.0
excel2json==1.0.1
excel2json-3==0.1.6
Flask==1.1.2
fonttools==4.37.1
Jinja2==2.11.3
joblib==1.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.24
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pymssql==2.2.5
PyMySQL==1.0.2
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
python-excel2json==0.2
pytz==2021.1
requests==2.28.1
six==1.15.0
SQLAlchemy==1.4.41
sqlparse==0.4.1
threadpoolctl==3.1.0
urllib3==1.26.11
virtualenv==20.16.5
xlrd==2.0.1
